+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Bump `envisim_utils`
+- Bump `envisim_utils`.
+- Added spatial_balance_energy measure.
 
 ## [0.4.0] - 2025-08-22
 ### Changed